pyproject.toml
[project]
dependencies = [
  "jupyterlab (>=4.4.9,<5.0.0)",
  "psycopg2-binary (>=2.9.11,<3.0.0)",
  "sqlalchemy (>=2.0.44,<3.0.0)",
  "pydantic (>=2.12.2,<3.0.0)",
  "pyarrow (>=21.0.0,<22.0.0)",
  "pandas (>=2.3.3,<3.0.0)",
  "plotly (>=6.3.1,<7.0.0)",
  "numpy (>=2.3.3,<3.0.0)",
  "scikit-learn (>=1.7.2,<2.0.0)",
  "python-dotenv (>=1.1.1,<2.0.0)",
  "pyspark>=4.0.1",
]
description = ""
name = "jupyter"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"
